Assigns operator based on group membership
When a chat has an exclusive group, the default group is not added.

When picking an operator the chat gets a list of operators available in each of
the groups the chat is tagged with. Each groups list is then combined and the
first available operator is picked from that list.