Removes dependency on Socket.IO socket instances
the chatlist middleware now can communicate with clients on other nodes.

This removes the ability to interact with socket instances directly, so instead of joining operators to chat specific rooms, this uses the operator identities to send chat messages to operator rooms.

This also removes the ability of the server to use socket.io callbacks to confirm that an operator's client successfully opened a chat. This will require an additional event emit from the client.